Modifications to allow use with compilers/run times that have c_int /= integer (e.g. g95).

+++ examples/cairo-basics-click.f90
@@ -50,7 +50,7 @@ module handlers
   integer(c_int) :: run_status = TRUE
   integer(c_int) :: boolresult
   logical :: boolevent
-  integer :: width, height
+  integer(kind=c_int) :: width, height
 
 
 contains
